Only display datasets by default
The CKAN behaviour is to only display dataset on the dataset search
page, but e.g. the organization pages displays all packages that belong
to an organization (including harvest sources).

This change makes sure that the filter for packages with the type
dataset is always applied if no such filter already exists.

This means that a plugin that implements it's own dataset type should
not be affected, if it is implemented correctly.