Create resource views on-the-fly in before_view
Unfortunately in CKAN 2.4 we don't yet have the IDataPusher interface
and therefore can't use the after_upload hook to do this in a nice way.
This solution now simply creates the default resource views before a
dataset is being viewed by a user, which could be a performance issue
but it actually doesn't take very long, especially once all the views
have been created and the function basically has nothing to do.