Create resource views when upload to DataPusher happened
In the past the resource views were created in the before_view of the
PackageControllerInterface, which meant that this code was called a lot
(i.e. on the overview page, the detail page of a dataset or resource
etc.) and therefore caused performance problems. With this change the
loading time of the /dataset page could be significantly improved.