Use separate fields for the suggest context
Unfortunately this is needed because the Solr suggester does not
properly handle context field values with dashes. They are somehow
treated as separate words and never match when queried.

Therefore I created new fields, where all dashes are stripped from the
values. These are only used for the suggestion context. In order for
this to work, the API *must* be called with dashes stripped as well.